+++ packages/electrode-archetype-react-app-dev/lib/dev-admin/log-parser.js
@@ -7,6 +7,7 @@ const { Levels } = require("./log-reader");
 const LogParse = /^(?:\u001b\[[0-9]+?m)?([a-z]+)(?:\u001b\[[0-9]+?m)?:(?: ([\s\S]*))?$/;
 // eslint-disable-next-line no-control-regex
 const FyiLogParse = /^(?:\u001b\[[0-9]+?m)?FYI ([a-z]+):(?:\u001b\[[0-9]+?m)?(?: ([\s\S]*))?$/;
+const UnhandledRejection = /([a-zA-Z]+): Unhandled rejection .*/;
 
 const FyiTag = ck`<yellow.inverse>[fyi]</> `;
 const BunyanTag = ck`<cyan.inverse>[byn]</> `;
@@ -19,12 +20,16 @@ const BunyanLevelLookup = {
   10: "silly"
 };
 const parsers = [
+  {
+    custom: (raw) => raw.match(UnhandledRejection) ? [raw, "error", raw] : undefined,
+    prefix: ""
+  },
   {regex: LogParse, prefix: ""},
   {regex: FyiLogParse, prefix: FyiTag}
 ];
 
 function parseRegex(raw, parser) {
-  const match = raw.match(parser.regex);
+  const match = parser.custom ? parser.custom(raw) : raw.match(parser.regex);
   if (!match) {
     return false;
   }