+++ packages/electrode-archetype-react-app-dev/test/spec/dev-admin/log-parser.spec.js
@@ -25,6 +25,20 @@ describe("log-parser", function() {
     expect(message).equal("An issue was detected in electrode error: Unable to load secrets file /secrets/ccm-secrets.json");
   });
 
+  it("should detect Unhandled rejection messages as an error when they are annotated 'info'", () => {
+    const raw = "info: Unhandled rejection ReferenceError: PropTypes is not defined";
+    const { level, message } = parse(raw);
+    expect(level).equal("error");
+    expect(message).equal(raw);
+  });
+
+  it("should detect Unhandled rejection messages as an error when they are annotated 'debug'", () => {
+    const raw = "debug: Unhandled rejection (rejection id: 3): TypeError: Cannot read property 'Electrode' of undefined";
+    const { level, message } = parse(raw);
+    expect(level).equal("error");
+    expect(message).equal(raw);
+  });
+
   it("should return correct level and message with badge for an FYI error", () => {
     const raw = "FYI error: Something unexpected happened";
     const { level, message } = parse(raw);